+++ newrelic/provider.go
@@ -4,13 +4,19 @@ import (
 	"fmt"
 	"log"
 
+	"github.com/hashicorp/terraform-plugin-sdk/httpclient"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/terraform"
+
+	"github.com/terraform-providers/terraform-provider-newrelic/version"
 )
+
+// TerraformProviderProductUserAgent string used to identify this provider in User Agent requests
+const TerraformProviderProductUserAgent = "terraform-provider-newrelic"
 
 // Provider represents a resource provider in Terraform
 func Provider() terraform.ResourceProvider {
-	return &schema.Provider{
+	provider := &schema.Provider{
 		Schema: map[string]*schema.Schema{
 			"api_key": {
 				Type:        schema.TypeString,
@@ -50,28 +56,38 @@ func Provider() terraform.ResourceProvider {
 			"newrelic_synthetics_monitor":         resourceNewRelicSyntheticsMonitor(),
 			"newrelic_synthetics_monitor_script":  resourceNewRelicSyntheticsMonitorScript(),
 		},
-
-		ConfigureFunc: providerConfigure,
 	}
+
+	provider.ConfigureFunc = func(d *schema.ResourceData) (interface{}, error) {
+		terraformVersion := provider.TerraformVersion
+		if terraformVersion == "" {
+			// Catch for versions < 0.12
+			terraformVersion = "0.11+compatible"
+		}
+		return providerConfigure(d, terraformVersion)
+	}
+
+	return provider
 }
 
-func providerConfigure(data *schema.ResourceData) (interface{}, error) {
+func providerConfigure(data *schema.ResourceData, terraformVersion string) (interface{}, error) {
 	config := Config{
-		APIKey: data.Get("api_key").(string),
-		APIURL: data.Get("api_url").(string),
+		APIKey:    data.Get("api_key").(string),
+		APIURL:    data.Get("api_url").(string),
+		userAgent: fmt.Sprintf("%s %s/%s", httpclient.TerraformUserAgent(terraformVersion), TerraformProviderProductUserAgent, version.ProviderVersion),
 	}
 	log.Println("[INFO] Initializing New Relic client")
 
 	client, err := config.Client()
 	if err != nil {
-		return nil, fmt.Errorf("Error initializing New Relic client: %s", err)
+		return nil, fmt.Errorf("error initializing New Relic client: %s", err)
 	}
 
 	log.Println("[INFO] Initializing New Relic Synthetics client")
 
 	clientSynthetics, err := config.ClientSynthetics()
 	if err != nil {
-		return nil, fmt.Errorf("Error initializing New Relic synthetics client: %s", err)
+		return nil, fmt.Errorf("error initializing New Relic synthetics client: %s", err)
 	}
 
 	infraConfig := Config{
@@ -82,7 +98,7 @@ func providerConfigure(data *schema.ResourceData) (interface{}, error) {
 
 	clientInfra, err := infraConfig.ClientInfra()
 	if err != nil {
-		return nil, fmt.Errorf("Error initializing New Relic Infra client: %s", err)
+		return nil, fmt.Errorf("error initializing New Relic Infra client: %s", err)
 	}
 
 	providerConfig := ProviderConfig{