+++ scripts/gofmtcheck.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 # Check gofmt
-echo "==> Checking that code complies with gofmt requirements..."
+#echo "==> Checking that code complies with gofmt requirements..."
 gofmt_files=$(gofmt -l `find . -name '*.go' | grep -v vendor`)
 if [[ -n ${gofmt_files} ]]; then
     echo 'gofmt needs running on the following files:'